Set up theme for variant A cookie banner
CCRC will be used for the cookie banner experiment, and will host variant A. Some changes needed to be made to the base theme, to pull it in and display it. This commit makes those changes. This includes adding a tag into the base file, and altering the base pixel size, as it was breaking the Design System component styling.